[precompile] generate nonce key if enable_aot_compile is enabled
As discussed offline with jamesjwu and aorenste in a precompile world,
it's actually not necessary to ensure we can safetely generate a unique
cache key since there is no implicit sharing. This PR adds a fallback so
in the case where we can't safetly generate a key for caching (eg. certain HOPs),
we still generate a random nonce key for precompile.

ghstack-source-id: cc66010f4ca4d228c94fcb315826d3be9a2778a5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/169244